+++ src/Products/MeetingLalouviere/tests/testMeetingItem.py
@@ -21,7 +21,7 @@ class testMeetingItem(MeetingLalouviereTestCase, mctmi):
     def _extraNeutralFields(self):
         """This method is made to be overrided by subplugins that added
         neutral fields to the MeetingItem schema."""
-        return ["followUp", "providedFollowUp"]
+        return ["followUp", "providedFollowUp", "interventions", "committeeTranscript"]
 
     def _get_developers_all_reviewers_groups(self):
         return [